+++ src/repo-name-arg-parser.ts
@@ -1,6 +1,0 @@
-import { CliStringArgParser } from '@carnesen/cli';
-
-export const repoNameArgParser = CliStringArgParser({
-  placeholder: '<repo name> | all',
-  required: true,
-});