+++ src/root/repos/index.ts
@@ -1,0 +1,8 @@
+import { CliBranch } from '@carnesen/cli';
+import { listCommand } from './list-command';
+import { cloneCommand } from './clone-command';
+
+export const reposBranch = CliBranch({
+	name: 'repos',
+	subcommands: [cloneCommand, listCommand],
+});